altera tipos de variáveis de trocas e comparações para long long int e configura locale para UTF-8

+++ src/algorithm.h
@@ -17,7 +17,7 @@ void mergeSort(jogador[], int, int);
 void countingSort(jogador[], int, int);
 void radixSort(jogador[], int);
 
-extern long int comparacoes;
-extern long int trocas;
+extern long long int comparacoes;
+extern long long int trocas;
 
 #endif
